requirements.txt
fastapi
uvicorn
langchain
google-generativeai
python-dotenv
jinja2
python-multipart